requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
procgen>=0.10.0
gym3>=0.3.0
tensorboard>=2.8.0
tqdm>=4.62.0
pyyaml>=5.4.0
umap-learn>=0.5.0
pytorch-grad-cam>=0.3.0
captum>=0.6.0  # for LRP
pillow>=8.0.0
jupyter>=1.0.0
ipykernel>=6.0.0 